The console runner includes a stub of a ListenerFactory with a non-default constructor to demonstrate the desired behavior of passing ListenerFactory constructor arguments across the AppDomain boundary.
As of this commit, tests against the main Fixie.Tests assembly do work, while tests against the Fixie.Samples assembly do not work.  This failure is due to the fact that the Fixie.Tests assembly has a direct reference to Fixie.Console.exe and therefore that assembly is already available at the moment we try to pass our concrete ExecutionSink across the AppDomain boundary when running the Fixie.Tests tests.  Fixie.Console.exe is not readily available to the child AppDomain when running the Fixie.Samples tests because it's not physically present in any place the normal assembly loader would look.  Although we successfully proxy across the boundary as an IExecutionSink, Activator.CreateInstance fails.  Activator.CreateInstance sees the object as Fixie.Console.ExecutionSink when Fixie.Console.exe was findable; it sees it as an opaque MarshalByRefObject when Fixie.Console.exe was not findable.