Remove support for deprecated TestAssembly subclassing.
The original intention of the TestAssembly class was to enable reuse of custom Convention classes across assemblies, but that concept is flawed.

First, as soon as a test assembly has a TestAssembly subclass, developers have to remember to add to it in order to truly execute newly defined convention classes, making it easy to silently fail to run intended tests.

Second, it was introduced at a time when most convention lifecycle hooks supported only lambda expressions. Now that lifecycle hooks also support class-accepting overloads, conventions are no longer the unit of reuse. Complex conventions can be expressed as small Convention subclasses that refer to other custom behavior classes, and those custom behavior classes are inherently reusable already.