When determining whether a generic test case MethodInfo has been sufficiently resolved to be invoked, use MethodInfo.ContainsGenericParameters instead of MethodInfo.IsGenericMethodDefinition. IsGenericMethodDefintion is insufficient in that it does not recursively inspect the surrounding type or type parameters for their own open generics.
From the Microsoft documentation on ContainsGenericParameters:

"To invoke a generic method, there must be no generic type definitions
or open constructed types in the type arguments of the method itself,
or in any enclosing types. Making this recursive determination can be
difficult. For convenience, and to reduce the chance of error, the
ContainsGenericParameters property provides a standard way to distinguish
between closed constructed methods, which can be invoked, and open
constructed methods, which cannot. If the ContainsGenericParameters
property returns true, the method cannot be invoked."

https://msdn.microsoft.com/en-us/library/system.reflection.methodinfo.containsgenericparameters(v=vs.85).aspx