Include the time spent in the Behavior chain into the Case Duration
Previously the Duration reported by a Case only included the duration that
it took to execute the case. It ignored the duration of all other build
chain behaviors that run as part of execution a Class.

With this change, we time the duration of the entire class chain and
adjust the case durations to also include the proportional duration of the
build chain execution.