For Fixie.Console and Fixie.Console.x86, *manually* set the project <OutputType> to "exe".
This is functionally equivalent to setting "Output type" to "Console Application"
on the Project Properties screen while avoiding a few Project Properties screen
bugs. Project Properties would have only defined the setting for the *active*
build configuration, forcing you to change the active build configuration at the
solution level in order to make the changes complete, and would have introduced
additional irrelevant XML tags whose default values are sufficient when omitted.

This manual change matches that of an example found in the documentation for the
"dotnet build" command.