The build script uses `dotnet test` to run all tests.
This works against the Fixie.Tests project with no special attention, because the `dotnet test` finds the Fixie.VisualStudio.TestAdapter.dll in that project's output folder. It's notable that this works even though the referenced Microsoft.VisualStudio.TestPlatform.ObjectModel.dll is an old version from before Visual Studio 2017.

For the Fixie.Samples project, however, the `--test-adapter-path` argument is required, since the adapter is not already present in the Fixie.Samples build output folder.

We expect `--test-adapter-path` will not be necessary when consuming Fixie as a NuGet package.