The Fixie.Tests project no longer incorrectly uses the RollForward capability:
This was included by mistake at the same time that RollForward was applied to Fixie.Console. It is appropriate to Fixie.Console because that defines a `dotnet` global tool, which needs to balance having a single build against the lowest supported target framework while also being "polite" in environments that lack that minimum supported version of the SDK. Fixie.Console deliberately uses a minimal and stable subset of the framework, so is considered safe enough to leverage RollForward. For Fixie.Tests, though, RollForward was hiding potential behavior differences across possible supported target frameworks, and could provoke surprising differences in behavior from one development environment to the next (e.g. local vs GitHub Actions).