Refuse to attempt debugger-attachment during the discovery phase, when LaunchProcessWithDebuggerAttached would be silently skipped.
Third-party test runners may or may not have the debugger attached to
the currently running Fixie.TestAdapter process during the discovery
phase. That is their choice and for some test frameworks that may even
be meaningful.

In Fixie's case, debugger-attached runs only make sense when the
secondary test assembly process we want to launch is being launched by
the IFrameworkHandle.LaunchProcessWithDebuggerAttached method. However,
there is no IFrameworkHandle instance during any Discovery phase, so by
here it is null.

If we were to attempt to run the debugger-attachment code during
discovery, where the IFrameworkHandle is null, the process launch
would be silently skipped with the use of `?.`, and the user would
experience errors while the interprocess communication between the
currently running Fixie.TestAdapter process and the *nonexistent*
secondary test assembly process inevitably failed.