Rephrase the adding of a key/value pair to a `ProcessStartInfo.Environment` dictionary, to get ahead of anticipated nullability warnings.
The intention is to change this project's target framework from netcoreapp3.1 to net6.0, but doing so would introduce a new warning (one treated rightly as an error), due to improved nullability markers added to the framework after netcoreapp3.1. The warning would have been:

> CS8620	Argument of type 'KeyValuePair<string, string>' cannot be used for parameter 'item' of type 'KeyValuePair<string, string?>' in 'void ICollection<KeyValuePair<string, string?>>.Add(KeyValuePair<string, string?> item)' due to differences in the nullability of reference types.

Since the framework will come to treat environment variable values as possibly-null, we need to be more clear that our values here are not null.

In our case, simply calling the overload that accepts the key and value separately avoids any risk around the upcoming framework change, so there will  be nothing to warn about.