Drop `[STAThread]` from entry points as a legacy default most appropriate in the .NET Framework / WinForms era:
The occurrence in `Fixie.Console` was never truly necessary as `dotnet fixie` never makes COM interop calls.

The occurrence in the injected `Main` methods for C# and F# test projects could possibly make COM interop calls, such as if the end user's own tests or system-under-test make such calls. Users who need STAThread semantics for COM interop are free to set it explicitly as demonstrated in the temporary custom `IExecution` body seen in this commit. The output from running this on Windows is:

> Running Fixie.Tests
>
> Entering CustomExecution STA Setup
>         Current Thread Apartment State: MTA
>         Attempting to set Apartment State to STA...
>         Current Thread Apartment State: STA
> Exiting CustomExecution STA Setup
> 149 passed, took 1.55 seconds