Apply MinVer to all projects using a GlobalPackageReference.
According to documentation, GlobalPackageReference silently takes care of the former explicit PrivateAssets and IncludeAssets values. From https://learn.microsoft.com/en-us/nuget/consume-packages/central-package-management#global-package-references we have:

> A global package reference is used to specify that a package will be used by every project in a repository. This includes packages that do versioning, extend your build, or any other packages that are needed by all projects. Global package references are added to the PackageReference item group with the following metadata:
>
> - IncludeAssets="Runtime;Build;Native;contentFiles;Analyzers"
>     This ensures that the package is only used as a development dependency and prevents any compile-time assembly references.
> - PrivateAssets="All"
>     This prevents global package references from being picked up by downstream dependencies.

Note that this technically *begins* the versioning of Fixie.Tests, but that is good as before this change it was getting the misleading default of 1.0.0.0, and now at least it versions accurately with the system under test.