Remove deprecated AppVeyor integration.
Users who wish to restore the integration can do so with a custom report equivalent to the one removed in his commit, though the plain ConsoleReport should be sufficient in this environment.

AppVeyor has never extended their test results API beyond their Windows images, which we take to be a vote of no confidence in the relevance of the feature. Fixie focuses on cross platform .NET target frameworks, so non-Windows builds are recommended for speed and compute resource reasons.