Drop deprecated interprocess communication between the test assembly process and the VSTest adapter process.
To limit the scope of this change to the interprocess communication mechanism and its immediate "footprint", the scope was to begin by deleting `PipeMessage.cs` and then remove anything that became an immediate compiler error or test failure until we could get back to a valid build with all passing tests.

As of this commit, there are remaining sections of dead code to be removed in subsequent commits, allowing this commit to represent the interprocess communciation mechanism itself should it or something like it need to be restored later.